Several enhancements:
- List activity shows progress
- ML label detection happens along with scraping
- All labels are persisted

+++ app/src/main/java/com/erraticduck/instagramcurate/cloud/InstagramService.kt
@@ -87,7 +87,7 @@ interface InstagramService {
 
     data class Edges<T>(
         val edges: List<Edge<T>>,
-        val count: Long?,
+        val count: Int?,    // Can be null with paged responses
         @SerializedName("page_info") val pageInfo: PageInfo?
     )
